Move static behavior out of Action
